Pass request_account_dir to request config
Previously, We compute and put the request_account_dir in into the
dehydrated config in dehydrated::certificate::request, and then we
re-compute the path in dehydrated_job_runner.rb for creating the account
if needed.

Now instead, we pass the request_account_dir we are giving to dehydrated
to the config so dehydrated_job_runner.rb can use it.